--- plugins/vulkan/pipeline/PipelineHelpers.cpp
@@ -1,11 +1,0 @@
-#include "PipelineHelpers.h"
-
-#include "PipelineBuilder.h"
-
-#include <Runtime/SystemPaths.h>
-#include <Log.h>
-
-namespace chai::gfx
-{
-
-} // namespace chai::gfx
